Rename endpoint "weather" to "current"
